# docs/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: Enclii Platform API
  description: |
    Open source DevOps platform API for deploying and managing containerized applications.

    ## Overview

    Enclii Switchyard is the control plane API that manages projects, services, deployments,
    domains, teams, and infrastructure. It provides a comprehensive REST API with 107+ endpoints
    covering the complete application lifecycle.

    ## Authentication

    Most endpoints require JWT authentication. There are two authentication modes:

    - **OIDC Mode** (Production): Uses Janua SSO with RS256 JWT validation
    - **Local Mode** (Development): Uses self-signed JWT with password-based auth

    Include the token in the `Authorization` header as `Bearer <token>`.

    ## Rate Limiting

    - Auth endpoints: 10 requests/minute (5 for register/login)
    - Standard endpoints: 100 requests/minute
    - Build endpoints: 20 requests/minute

    ## Environments

    Services can be deployed to multiple environments:
    - `dev` - Development (default)
    - `staging` - Pre-production testing
    - `prod` - Production

    ## WebSocket Endpoints

    Real-time log streaming is available via WebSocket:
    - `ws://api.enclii.dev/v1/services/{id}/logs/stream` - Service logs
    - `ws://api.enclii.dev/v1/deployments/{id}/logs/stream` - Deployment logs
    - `ws://api.enclii.dev/v1/services/{id}/builds/{build_id}/logs/stream` - Build logs

  version: 1.0.0
  contact:
    name: MADFAM Engineering
...
servers:
...
paths:
...
  /dashboard/stats:
    get:
      summary: Dashboard statistics
      description: |
        Returns aggregated statistics for the dashboard including healthy services count,
        deployments today, active projects, and recent activity. Requires authentication
        (same session as other `/v1` protected routes).
      tags: [health]
      operationId: getDashboardStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'

  # ============================================
  # AUTHENTICATION ENDPOINTS
  # ============================================
...
  /projects:
    get:
      summary: List projects
      description: Get all projects accessible to the authenticated user.
      tags: [projects]
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      summary: Create project
      description: Create a new project. Requires admin role.
      tags: [projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '409':
          description: Project slug already exists
...
  /projects/{slug}/services:
    get:
      summary: List services
      description: Get all services in a project.
      tags: [services]
      operationId: listServices
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service list
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
    post:
      summary: Create service
      description: Create a new service in a project.
      tags: [services]
      operationId: createService
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or OIDC callback
  parameters:
    limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 50
        maximum: 100
    offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        default: 0
...
  schemas:
...
    DashboardStats:
      type: object
      properties:
        healthy_services:
          type: integer
        deployments_today:
          type: integer
        active_projects:
          type: integer
        recent_activity:
          type: array
          items:
            $ref: '#/components/schemas/Activity'

    # ===== Auth =====
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 2
          maxLength: 50

    # ===== Environments =====
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        git_repo:
          type: string
          description: GitHub HTTPS URL
        git_branch:
          type: string
        app_path:
          type: string
          description: Path in monorepo
        auto_deploy:
          type: boolean
        auto_deploy_branch:
          type: string
        auto_deploy_env:
          type: string
        build_config:
          $ref: '#/components/schemas/BuildConfig'
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP response headers injected via nginx ingress annotations
          example:
            Cross-Origin-Opener-Policy: same-origin
            Cross-Origin-Embedder-Policy: require-corp
        status:
          type: string
          enum: [running, stopped, building]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BuildConfig:
      type: object
      properties:
        type:
          type: string
          enum: [auto, dockerfile, buildpack]
        dockerfile:
          type: string
        buildpack:
          type: string
    CreateServiceRequest:
      type: object
      required:
        - name
        - git_repo
      properties:
        name:
          type: string
        git_repo:
          type: string
        app_path:
          type: string
        auto_deploy:
          type: boolean
          default: true
        auto_deploy_branch:
          type: string
          default: main
        build_config:
          $ref: '#/components/schemas/BuildConfig'
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP response headers injected via nginx ingress annotations
...
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
          format: uuid
        actor_id:
          type: string
          format: uuid
        actor_name:
          type: string
        details:
          type: object
        created_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []